In setup_links.sh check the file exists before trying to move.

+++ setup_links.sh
@@ -11,7 +11,10 @@ else
   then
     echo "skipping bashrc..."
   else
-    mv ~/.bashrc ~/.bashrc_env_backup
+    if [ -f ~/.bashrc ]
+    then
+      mv ~/.bashrc ~/.bashrc_env_backup
+    fi
     echo "Linking bashrc"
     ln -s ~/tmp/dotfiles/bashrc ~/.bashrc
     cp ~/tmp/dotfiles/env_settings ~/.env_settings
@@ -26,7 +29,10 @@ if [ -f ~/.gemrc_env_backup ]
 then
   echo "skipping gemrc..."
 else
-  mv ~/.gemrc ~/.gemrc_env_backup
+  if [ -f ~/.gemrc ]
+  then
+    mv ~/.gemrc ~/.gemrc_env_backup
+  fi
   echo "Linking gemrc"
   ln -s ~/tmp/dotfiles/gemrc ~/.gemrc
 fi
@@ -35,7 +41,10 @@ if [ -f ~/.gitconfig_env_backup ]
 then
   echo "skipping gitconfig..."
 else
-  mv ~/.gitconfig ~/.gitconfig_env_backup
+  if [ -f ~/.gitconfig ]
+  then
+    mv ~/.gitconfig ~/.gitconfig_env_backup
+  fi
   echo "Linking gitconfig"
   ln -s ~/tmp/dotfiles/gitconfig ~/.gitconfig
 fi
@@ -44,7 +53,10 @@ if [ -f ~/.githelpers_env_backup ]
 then
   echo "skipping githelpers..."
 else
-  mv ~/.githelpers ~/.githelpers_env_backup
+  if [ -f ~/.githelpers ]
+  then
+    mv ~/.githelpers ~/.githelpers_env_backup
+  fi
   echo "Linking githelpers"
   ln -s ~/tmp/dotfiles/githelpers ~/.githelpers
 fi
@@ -53,7 +65,10 @@ if [ -f ~/.screenrc_env_backup ]
 then
   echo "skipping screenrc..."
 else
-  mv ~/.screenrc ~/.screenrc_env_backup
+  if [ -f ~/.screenrc ]
+  then
+    mv ~/.screenrc ~/.screenrc_env_backup
+  fi  
   echo "Linking screenrc"
   ln -s ~/tmp/dotfiles/screenrc ~/.screenrc
 fi
@@ -62,7 +77,10 @@ if [ -f ~/.tmux_env_backup ]
 then
   echo "skipping tmux..."
 else
-  mv ~/.tmux.conf ~/.tmux_env_backup
+  if [ -f ~/.tmux.conf ]
+  then
+    mv ~/.tmux.conf ~/.tmux_env_backup
+  fi
   echo "Linking tmux.conf"
   ln -s ~/tmp/dotfiles/tmux.conf ~/.tmux.conf
 fi
@@ -71,7 +89,10 @@ if [ -f ~/.vimrc_env_backup ]
 then
   echo "skipping vimrc..."
 else
-  mv ~/.vimrc ~/.vimrc_env_backup
+  if [ -f ~/.vimrc ]
+  then
+    mv ~/.vimrc ~/.vimrc_env_backup
+  fi  
   echo "Linking vimrc"
   ln -s ~/tmp/dotfiles/vimrc ~/.vimrc
 fi
@@ -80,7 +101,10 @@ if [ -f ~/.vim_env_backup ]
 then
   echo "skipping vim..."
 else
-  mv ~/.vim ~/.vim_env_backup
+  if [ -f ~/.vim ]
+  then
+    mv ~/.vim ~/.vim_env_backup
+  fi
   echo "Linking vim"
   ln -s ~/tmp/dotfiles/vim ~/.vim
 fi
@@ -89,7 +113,10 @@ if [ -f ~/.psqlrc_env_backup ]
 then
   echo "skipping psqlrc..."
 else
-  mv ~/.psqlrc ~/.psqlrc_env_backup
+  if [ -f ~/.psqlrc ]
+  then
+    mv ~/.psqlrc ~/.psqlrc_env_backup
+  fi
   echo "Linking psqlrc"
   ln -s ~/tmp/dotfiles/psqlrc ~/.psqlrc
 fi
@@ -98,7 +125,10 @@ if [ -f ~/.zshrc_env_backup ]
 then
   echo "skipping zshrc..."
 else
-  mv ~/.zshrc ~/.zshrc_env_backup
+  if [ -f ~/.zshrc ]
+  then
+    mv ~/.zshrc ~/.zshrc_env_backup
+  fi
   echo "Linking zshrc"
   ln -s ~/tmp/dotfiles/zshrc_linux ~/.zshrc
 fi
